Add create SAR e2e test
