+++ package.json
@@ -28,6 +28,7 @@
     "@enterprise-cmcs/serverless-waf-plugin": "^1.4.0"
   },
   "devDependencies": {
+    "@faker-js/faker": "^8.4.1",
     "@playwright/test": "^1.44.0",
     "@serverless/compose": "^1.3.0",
     "@stratiformdigital/serverless-stage-destroyer": "^2.0.0",
@@ -37,6 +38,7 @@
     "@types/yargs": "^15.0.10",
     "@typescript-eslint/eslint-plugin": "5.18.0",
     "@typescript-eslint/parser": "5.18.0",
+    "amazon-user-pool-srp-client": "^1.0.4",
     "dotenv": "^8.2.0",
     "eslint": "^8.33.0",
     "eslint-config-airbnb": "^19.0.4",
