+++ tests/seeds/fixtures/index.ts
@@ -1,0 +1,3 @@
+export * from "./banner";
+export * from "./semi-annual-report";
+export * from "./work-plan";
