+++ tests/seeds/types.ts
@@ -1,4 +1,7 @@
-import { ReportFieldData, ReportMetadataShape } from "../utils/types";
+import {
+  ReportFieldData,
+  ReportMetadataShape,
+} from "../../services/app-api/utils/types";
 
 export type AwsHeaders = {
   "Content-Type"?: string;
