+++ tests/tsconfig.json
@@ -1,0 +1,15 @@
+{
+  "compilerOptions": {
+    "strict": true,
+    "outDir": ".build",
+    "lib": ["ES2022"],
+    "module": "node16",
+    "target": "ES2022",
+  },
+  "include": [
+    "seeds"
+  ],
+  "files": [
+    "seeds/custom.d.ts"
+  ],
+}
